refactor(coach): REFACTOR M001 — split watchers into per-component modules
Each module corresponds to one `produce` entry on the
drive-state-machine wagon manifest:
- event_queue.py: CoachEventQueue, NATURAL_KEY, REPLAY_BEHAVIOR
- runtime_watcher.py: RuntimeWatcher (`commons:coach:runtime-event-watcher`)
- git_watcher.py: GitWatcher + parse_commit_trailers (`commons:coach:git-event-watcher`)
- liveness_checker.py: LivenessChecker (`commons:coach:liveness-checker`)

watchers.py is preserved as a thin re-export so downstream consumers
(#L1 observer, #J6 resume runner) keep a single import surface.
